# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent-Native Middleware API",
    "description": "## Agent-Native Middleware API\n\nHeadless infrastructure for the B2A (Business-to-Agent) economy. This API provides four core services that give AI agents the 'hands' and 'eyes' to manipulate the digital and physical world.\n\n### Service Pillars\n\n- **IoT Protocol Bridge** (`/v1/iot`) \u2014 Secure, ACL-enforced protocol translation for IoT devices (MQTT, CoAP, Zigbee, etc.)\n- **Autonomous Product Manager** (`/v1/telemetry`) \u2014 Ingest telemetry, detect anomalies, auto-generate pull requests to fix bugs\n- **Programmatic Media Engine** (`/v1/media`) \u2014 Video ingestion, viral hook detection, reframing, captioning, and cross-platform distribution\n- **Agent Communications** (`/v1/comms`) \u2014 Agent registration, structured messaging, capability discovery, and swarm task handoffs\n\n### Authentication\n\nAll endpoints require an API key passed via the `X-API-Key` header.\n\n### For Agents\n\nThis API is designed for programmatic consumption. See `/llm.txt` for LLM-optimized documentation and `/openapi.json` for the full spec.",
    "contact": {
      "name": "Agent-Native Middleware",
      "email": "api@yourdomain.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://api.yourdomain.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
...
    }
  ],
  "paths": {
...
    "/v1/security/scans": {
      "get": {
        "tags": [
          "Red Team Security Swarm"
        ],
        "summary": "List all security scans",
        "description": "Returns all historical scan reports, newest first.",
        "operationId": "list_scans_v1_security_scans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "post": {
        "tags": [
          "Red Team Security Swarm"
        ],
        "summary": "Launch Red Team scan",
        "description": "Deploy the security swarm to attack specified services with selected attack categories. Returns immediately with scan ID; the swarm runs asynchronously. Default: full scan of all services with all vectors.",
        "operationId": "launch_scan_v1_security_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/security/scans/{scan_id}": {
      "get": {
        "tags": [
          "Red Team Security Swarm"
        ],
        "summary": "Get scan report",
        "description": "Returns the full security scan report including all discovered vulnerabilities, severity breakdown, security score, and prioritized remediation recommendations.",
        "operationId": "get_scan_report_v1_security_scans__scan_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/security/scans/{scan_id}/vulnerabilities": {
      "get": {
        "tags": [
          "Red Team Security Swarm"
        ],
        "summary": "Get vulnerabilities from a scan",
        "description": "Returns just the vulnerabilities from a scan, optionally filtered by severity. Machine-readable format for the Autonomous PM to auto-generate fix PRs.",
        "operationId": "get_vulnerabilities_v1_security_scans__scan_id__vulnerabilities_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/app__schemas__red_team__Severity"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VulnerabilityListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/security/scans/quick": {
      "post": {
        "tags": [
          "Red Team Security Swarm"
        ],
        "summary": "Quick security check",
        "description": "Runs a fast, surface-level scan focused on CRITICAL and HIGH severity vectors only (ACL bypass, auth probes, privilege escalation). Returns the full report synchronously \u2014 use for CI/CD gates.",
        "operationId": "quick_scan_v1_security_scans_quick_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttackCategory": {
        "type": "string",
        "enum": [
          "acl_bypass",
          "auth_probe",
          "injection",
          "rate_limit_evasion",
          "privilege_escalation",
          "schema_abuse",
          "enumeration",
          "replay",
          "dos_pattern"
        ],
        "title": "AttackCategory",
        "description": "Categories of automated attack vectors."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RemediationStatus": {
        "type": "string",
        "enum": [
          "open",
          "acknowledged",
          "patched",
          "accepted_risk"
        ],
        "title": "RemediationStatus"
      },
      "ScanListResponse": {
        "properties": {
          "scans": {
            "items": {
              "$ref": "#/components/schemas/ScanReport"
            },
            "type": "array",
            "title": "Scans"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "scans",
          "total"
        ],
        "title": "ScanListResponse"
      },
      "ScanReport": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "target_services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Services"
          },
          "attack_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attack Categories"
          },
          "intensity": {
            "type": "string",
            "title": "Intensity"
          },
          "total_tests_run": {
            "type": "integer",
            "title": "Total Tests Run"
          },
          "total_passed": {
            "type": "integer",
            "title": "Total Passed"
          },
          "total_failed": {
            "type": "integer",
            "title": "Total Failed"
          },
          "vulnerabilities_found": {
            "type": "integer",
            "title": "Vulnerabilities Found"
          },
          "severity_breakdown": {
            "additionalProperties": true,
            "type": "object",
            "title": "Severity Breakdown",
            "description": "Count per severity level: {critical: N, high: N, ...}"
          },
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/Vulnerability"
            },
            "type": "array",
            "title": "Vulnerabilities"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "Prioritized remediation recommendations."
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "description": "Security score: 100 = fortress, 0 = DJI Romo vacuum."
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "status",
          "started_at",
          "target_services",
          "attack_categories",
          "intensity",
          "total_tests_run",
          "total_passed",
          "total_failed",
          "vulnerabilities_found",
          "score"
        ],
        "title": "ScanReport",
        "description": "Full scan report."
      },
      "ScanRequest": {
        "properties": {
          "target_services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Services",
            "description": "Which service pillars to attack. Default: all.",
            "default": [
              "iot",
              "telemetry",
              "media",
              "comms",
              "factory"
            ]
          },
          "attack_categories": {
            "items": {
              "$ref": "#/components/schemas/AttackCategory"
            },
            "type": "array",
            "title": "Attack Categories",
            "description": "Which attack vectors to deploy. Default: all.",
            "default": [
              "acl_bypass",
              "auth_probe",
              "injection",
              "rate_limit_evasion",
              "privilege_escalation",
              "schema_abuse",
              "enumeration",
              "replay",
              "dos_pattern"
            ]
          },
          "intensity": {
            "type": "string",
            "title": "Intensity",
            "description": "Scan intensity: 'quick' (surface-level), 'standard' (thorough), 'aggressive' (full fuzzing).",
            "default": "standard"
          },
          "auto_remediate": {
            "type": "boolean",
            "title": "Auto Remediate",
            "description": "If true, the swarm will attempt to auto-generate patches for discovered vulnerabilities.",
            "default": false
          }
        },
        "type": "object",
        "title": "ScanRequest",
        "description": "Initiate a Red Team scan."
      },
      "ScanResponse": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "target_services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Services"
          },
          "attack_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attack Categories"
          },
          "intensity": {
            "type": "string",
            "title": "Intensity"
          },
          "estimated_duration_seconds": {
            "type": "integer",
            "title": "Estimated Duration Seconds"
          },
          "total_attack_vectors": {
            "type": "integer",
            "title": "Total Attack Vectors"
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "status",
          "target_services",
          "attack_categories",
          "intensity",
          "estimated_duration_seconds",
          "total_attack_vectors"
        ],
        "title": "ScanResponse",
        "description": "Response after initiating a scan."
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed"
        ],
        "title": "ScanStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Vulnerability": {
        "properties": {
          "vuln_id": {
            "type": "string",
            "title": "Vuln Id"
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "category": {
            "$ref": "#/components/schemas/AttackCategory"
          },
          "severity": {
            "$ref": "#/components/schemas/app__schemas__red_team__Severity"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short description of the finding."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Detailed technical explanation including attack vector and impact."
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint",
            "description": "The affected API path."
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "HTTP method (GET, POST, etc.)."
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "description": "Proof-of-concept: request payload, response status, headers, etc."
          },
          "remediation": {
            "type": "string",
            "title": "Remediation",
            "description": "Recommended fix. Agents read this to auto-generate patches."
          },
          "remediation_status": {
            "$ref": "#/components/schemas/RemediationStatus",
            "default": "open"
          },
          "cwe_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwe Id",
            "description": "Common Weakness Enumeration ID (e.g. CWE-285 for improper authorization)."
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          }
        },
        "type": "object",
        "required": [
          "vuln_id",
          "scan_id",
          "category",
          "severity",
          "title",
          "description",
          "endpoint",
          "method",
          "remediation",
          "discovered_at"
        ],
        "title": "Vulnerability",
        "description": "A discovered security vulnerability."
      },
      "VulnerabilityListResponse": {
        "properties": {
          "vulnerabilities": {
            "items": {
              "$ref": "#/components/schemas/Vulnerability"
            },
            "type": "array",
            "title": "Vulnerabilities"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "critical_count": {
            "type": "integer",
            "title": "Critical Count"
          },
          "high_count": {
            "type": "integer",
            "title": "High Count"
          }
        },
        "type": "object",
        "required": [
          "vulnerabilities",
          "total",
          "critical_count",
          "high_count"
        ],
        "title": "VulnerabilityListResponse"
      },
      "app__schemas__red_team__Severity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low",
          "info"
        ],
        "title": "Severity"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API key for agent authentication. Pass in the X-API-Key header.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}